Issue #34: Port to python 3
Ported everything to python 3. Synopsis:

* `map(foo)` changed to `list(map(foo))`
* In packages, `import module` changed to `import .module` (absolute path)
* Convert to int explicitly: `int(int1/int2)` instead of just `int1/int2`

+++ utilities.py
@@ -40,16 +40,16 @@ class Counter:
 
 def split_tiled_image(image, tile_size, colorkey=None):
 	"""Splits a tiled image into single tiles in a return_list."""
-	
+
 	image_size = image.get_size()
 	tmp_surface = pygame.Surface(tile_size)
 	tmp_surface.set_colorkey(colorkey)
 	return_list = []
 
 	# For tilenr in y axis...
-	for y in range(image_size[1]/tile_size[1]):
+	for y in range(int(image_size[1]/tile_size[1])):
 		# For tilenr in x axis...
-		for x in range(image_size[0]/tile_size[0]):
+		for x in range(int(image_size[0]/tile_size[0])):
 			# ...blit the cutout at (tilenr_x*tilesize_x, tilenr_y*tilesize_y) on tmp_surface
 			tmp_surface.blit(image, (0, 0), ((x*tile_size[0], y*tile_size[1]), tile_size))
 			# Append a copy of tmp_surface to the return_list
@@ -57,4 +57,3 @@ def split_tiled_image(image, tile_size, colorkey=None):
 			# Fill the tmp_surface black again
 			tmp_surface.fill((0, 0, 0))
 	return return_list
-
